refactor: scrollToLast URL 파라미터 로직을 useScrollToLast 훅으로 분리
